cloning works, but release no longer works. I believe what's happening is that they share the manager objects.
Instead, each manager needs to be created, and then fetch the object of interest to read. That means that each thread, when released, will have their own local socket that won't be closed on clone.

More expensive? Yes. But may solve the weird proxy object failures (the only known error for this approach right now).