Holy cow, that was a rough go of it. Fixed and runs MH, but I haven't run through tests yet.
For a number of reasons, it's hard to track forks, and there are a lot of subtle bugs that can come up.

It's not tremendously fast when you run only a single sample, which makes the current MCMC implementation kind of silly for what could potentially be done with fork replays.

This branch is currently dependent on Neeraj's hmc branch. I might temporarily commit his work so other can run this branch, then i'll remove the mcmc files.