Get rid of Poutine._prepare_site
As far as I can tell this function is unnecessary and everything currently done in the reverse pass through the Pyro stack can be done in the normal pass. This is because poutines are always composed in a sensible order, namely those that do preprocessing come before those that do actual work. In my opinion composition in the other direction would be weird. Still, I may have missed something, but the poutine tests pass without modification.