Remove most poutines from the stack
This commit removes the need to register most poutines on the Pyro stack. The logic was rewritten in message-passing style using greenlets which alleviates the need for global state. The only exceptions are scaling and independence poutines, which are still allocated on the Pyro stack since they must be added dynamically during execution. This is not a fundamental limitation and I expect to remove those poutines from the stack in subsequent commits.