Move raising NonlocalExit to a continuation in the message
This commit adds a continuation field to the message and modifies EscapeMessenger to raise the NonlocalExit exception in the continuation. The purpose is to allow the remaining components of the stack to finish postprocessing before the computation is aborted. This is because in case of an escape the messengers below Escape on the stack are expected to complete their operations while the ones above are not. On the other hand the messengers above on the stack are not expected to take effect, which I find rather confusing.