New features for Importance class
Added methods for the Importance class that are often used with importance sampling methods. Useful on their own, and for future features such as Sequential Monte Carlo and Annealed Importance Sampling.
In all cases they are computed in a numerically stable manner using the log scale.
	1	Estimate the normalizing constant of the target distribution. (this assumes that the log_weights are computed using the normalized density of the guide, otherwise it is an estimator of the ratio of normalising constants). This is used at each stage in Sequential Monte Carlo to construct marginal likelihood estimators, likewise for Annealed Importance Sampling.
	2	Computing the normalized sample weights.
	3	Computing the (Importance Sampling) effective sample size (which is different to the one used for MCMC). This is useful as a diagnostic for the efficacy of an importance sampling density. It is also used in Sequential Monte Carlo algorithms to determine resampling times.
For further background and an example of where such things are used, see for example: https://www.stats.ox.ac.uk/~doucet/delmoral_doucet_jasra_sequentialmontecarlosamplersJRSSB.pdf
which discusses these aspects in the context of SMC Samplers.