update intro tutorial (part ii) (#1524)
* update intro tutorial (part ii)

Addresses issue #1479: https://github.com/uber/pyro/issues/1479

This is a start of a reworked tutorial.  The original tutorial was difficult to follow, quite complex, and had some errors (particularly regarding the posterior distribution for the example). References to marginal distributions removed as discussed. Visualizations improved, and now it shows how SVI recovers something close to the true posterior. 

Also, importance sampling removed as this will be another tutorial. I have also removed comments regarding using the guide for importance sampling; this may not be a good recommendation. The way KL divergence is minimised with SVI will not cover tails correctly and variance can become quite large (expectation propagation does not have this problem, as the KL is minimised in other direction). 

Anyways, this is something I made mainly for myself and colleagues, but you may wish to merge it and work from here.

* update intro_part_ii.ipynb

Additional cleanup and editing.

* update intro_part_ii.ipynb after review

Addressed review comments by eb8680.

* nits