Remove uses of `@torch.jit.script`
This changes uses of `@torch.jit.script` to
`@torch.jit._script_if_tracing`. The TorchScript compiler has a lot of
lazy static initialization, so including these decorators will impact
the time to execute `import pyro` for everyone, not just jit users.
https://github.com/pytorch/pytorch/pull/34935 added a workaround for
this so that if a function has this lazy script decorator, it can still
be used to handle control flow in tracing.