Additional use of `self.dim`
We need to take `self.dim` into account at the end of `_call`, `_inverse`, and `log_abs_det_jacobian`.

As a clarification, the changes to `log_abs_det_jacobian` are needed to perform an "elementwise" summation, where the elements are the samples/batches. This ensures that the transform works with things like `TransformedDistribution` that use the transform's `event_dim` to sum log probabilities from different transformations.