Update __init__.py (#2555)
Fix 1. The z in the guide is sampled from the variational approximation of the true posterior z|y,t,x not t |y,t,x
Fix 2. using p for both the true and the variational approximation is poor notation and inconsistent with 
standard practice and the original paper. replacing p with q, to differentiate between the true and the 
variational distribution.