Allow registering of custom exception handlers for potential_fn computations (#3168)
* [RFC] Allow registering excpetion handlers for potential function computations

In some cases evaluation the potential funciton may result in numerical issues. Currently the code hard-codes the handling of a RuntimeError raised when matrices are (numerically) singular. This PR adds the ability to register custom exception handlers. This allows other code depending on pyro to register custom exception handlers without having to modify core pyro code.

There are some other places in which `potential_fn` is called that could benefit from being guarded by these handlers (one is `HMC._find_reasonable_step_size`). I'm not sure what the right thing to do there is when encountering numerical isssues, but happy to add this in as needed.

* Fix typing lint, typos.

* Warn instead of raise, fix typing import.

* Make isort happy (hopefully)

* Check for instance rather than type equality in _handle_torch_singular

* Handle numerical issues also in HMC._find_reasonable_step_size

* isort once more

* Fix black format